modularize chain data
