+++ src/client/mod.rs
@@ -18,7 +18,7 @@ use crate::{
 use crate::mock::MockRpcApi;
 
 pub mod accounts;
-pub mod chain_state;
+pub mod chain_data;
 pub mod notes;
 
 // CONSTANTS
