+++ src/store/mod.rs
@@ -9,7 +9,6 @@ use crypto::{
     Word,
 };
 
-use objects::ChainMmr;
 use objects::{
     accounts::{Account, AccountCode, AccountId, AccountStorage, AccountStub, AccountVault},
     assembly::{AstSerdeOptions, ModuleAst},
@@ -19,6 +18,7 @@ use objects::{
 };
 use rusqlite::{params, Connection, Transaction};
 
+pub mod chain_data;
 mod migrations;
 
 // TYPES
@@ -56,12 +56,6 @@ type SerializedAccountCodeParts = (String, String, String);
 
 type SerializedAccountStorageData = (String, String);
 type SerializedAccountStorageParts = (String, String);
-
-type SerializedBlockHeaderData = (i64, String, String, String, String);
-type SerializedBlockHeaderParts = (i64, String, String, String, String);
-
-type SerializedChainMmrNodeData = String;
-type SerializedChainMmrNodeParts = (i64, String);
 
 // CLIENT STORE
 // ================================================================================================
@@ -488,71 +482,6 @@ impl Store {
         }
         Ok(())
     }
-
-    // CHAIN DATA
-    // --------------------------------------------------------------------------------------------
-    pub fn insert_block_header(&mut self, block_header: BlockHeader) -> Result<(), StoreError> {
-        let (block_num, header, notes_root, sub_hash, chain_mmr) =
-            serialize_block_header(block_header)?;
-
-        const QUERY: &str = "\
-        INSERT INTO block_headers
-            (block_num, header, notes_root, sub_hash, chain_mmr)
-         VALUES (?, ?, ?, ?, ?)";
-
-        println!("inserting block header {}", block_num);
-        self.db
-            .execute(
-                QUERY,
-                params![block_num, header, notes_root, sub_hash, chain_mmr],
-            )
-            .map_err(StoreError::QueryError)
-            .map(|_| ())
-    }
-
-    pub fn get_block_header_by_num(&self, block_number: u32) -> Result<BlockHeader, StoreError> {
-        const QUERY: &str = "SELECT block_num, header, notes_root, sub_hash, chain_mmr FROM block_headers WHERE block_num = ?";
-        println!("getting block header {}", block_number as u64 as i64);
-        self.db
-            .prepare(QUERY)
-            .map_err(StoreError::QueryError)?
-            .query_map(params![block_number as i64], parse_block_headers_columns)
-            .map_err(StoreError::QueryError)?
-            .map(|result| {
-                result
-                    .map_err(StoreError::ColumnParsingError)
-                    .and_then(parse_block_header)
-            })
-            .next()
-            .ok_or(StoreError::BlockHeaderNotFound(block_number))?
-    }
-
-    pub fn insert_chain_mmr_node(&mut self, chain_mmr: ChainMmr) -> Result<(), StoreError> {
-        let node = serialize_chain_mmr(chain_mmr)?;
-
-        const QUERY: &str = "INSERT INTO chain_mmr_nodes (node) VALUES (?)";
-
-        self.db
-            .execute(QUERY, params![node])
-            .map_err(StoreError::QueryError)
-            .map(|_| ())
-    }
-
-    pub fn get_chain_mmr_hash_by_id(&self, id: u64) -> Result<ChainMmr, StoreError> {
-        const QUERY: &str = "SELECT id, node FROM chain_mmr_nodes WHERE id = ?";
-        self.db
-            .prepare(QUERY)
-            .map_err(StoreError::QueryError)?
-            .query_map(params![id as i64], parse_chain_mmr_nodes_columns)
-            .map_err(StoreError::QueryError)?
-            .map(|result| {
-                result
-                    .map_err(StoreError::ColumnParsingError)
-                    .and_then(parse_chain_mmr_nodes)
-            })
-            .next()
-            .ok_or(StoreError::ChainMmrNodeNotFound(id))?
-    }
 }
 
 // DATABASE AUTH INFO
@@ -912,61 +841,6 @@ fn serialize_input_note(
     ))
 }
 
-fn serialize_block_header(
-    block_header: BlockHeader,
-) -> Result<SerializedBlockHeaderData, StoreError> {
-    let block_num: u64 = block_header.block_num().into();
-    let header =
-        serde_json::to_string(&block_header).map_err(StoreError::InputSerializationError)?;
-    let notes_root = serde_json::to_string(&block_header.note_root())
-        .map_err(StoreError::InputSerializationError)?;
-    let sub_hash = serde_json::to_string(&block_header.sub_hash())
-        .map_err(StoreError::InputSerializationError)?;
-    let chain_mmr = serde_json::to_string(&block_header.chain_root())
-        .map_err(StoreError::InputSerializationError)?;
-
-    Ok((block_num as i64, header, notes_root, sub_hash, chain_mmr))
-}
-
-fn parse_block_headers_columns(
-    row: &rusqlite::Row<'_>,
-) -> Result<SerializedBlockHeaderParts, rusqlite::Error> {
-    let block_num: i64 = row.get(0)?;
-    let header: String = row.get(1)?;
-    let notes_root: String = row.get(2)?;
-    let sub_hash: String = row.get(3)?;
-    let chain_mmr: String = row.get(4)?;
-    Ok((block_num, header, notes_root, sub_hash, chain_mmr))
-}
-
-fn parse_block_header(
-    serialized_block_header_parts: SerializedBlockHeaderParts,
-) -> Result<BlockHeader, StoreError> {
-    let (_, header, _, _, _) = serialized_block_header_parts;
-
-    serde_json::from_str(&header).map_err(StoreError::JsonDataDeserializationError)
-}
-
-fn serialize_chain_mmr(chain_mmr: ChainMmr) -> Result<SerializedChainMmrNodeData, StoreError> {
-    serde_json::to_string(&chain_mmr).map_err(StoreError::InputSerializationError)
-}
-
-fn parse_chain_mmr_nodes_columns(
-    row: &rusqlite::Row<'_>,
-) -> Result<SerializedChainMmrNodeParts, rusqlite::Error> {
-    let id = row.get(0)?;
-    let node = row.get(1)?;
-    Ok((id, node))
-}
-
-fn parse_chain_mmr_nodes(
-    serialized_chain_mmr_node_parts: SerializedChainMmrNodeParts,
-) -> Result<ChainMmr, StoreError> {
-    let (_, node) = serialized_chain_mmr_node_parts;
-
-    serde_json::from_str(&node).map_err(StoreError::JsonDataDeserializationError)
-}
-
 // TESTS
 // ================================================================================================
 
@@ -995,7 +869,7 @@ pub mod tests {
         temp_file
     }
 
-    fn create_test_store() -> Store {
+    pub fn create_test_store() -> Store {
         let temp_file = create_test_store_path();
         let mut db = Connection::open(temp_file).unwrap();
         migrations::update_to_latest(&mut db).unwrap();
